testing/web-platform/tests/atomics/dedicated-worker-agent/wait/no-spurious-wakeup.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/workerhelper.js></script>
<script>

/*---
description: >
  Test that Atomics.wait actually waits and does not spuriously wake
  up when the memory value is changed.
includes: [atomicsHelper.js]
---*/

window.addEventListener("load", function(e) {
  async_test(function(t) {
		var worker_code = `
			onmessage = function(e) {
					var ia = e.data[0];
					var then = Date.now();
					Atomics.wait(ia, 0, 0);
					var diff = Date.now() - then;
					postMessage([diff]);
			}
		`;
		var onmessage = this.step_func(function(e) {
      var ret = e.data[0];
      assert_true(Math.abs(ret) <= 1000);
      t.done();
    });

		var worker = WorkerHelper.createFromString(worker_code, onmessage);
    var ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    worker.postMessage([ia]);
    setTimeout(function() {
      // Change the value, should not wake the agent.
      Atomics.store(ia, 0, 1);
      // Wake up worker.
      setTimeout(() => Atomics.wake(ia, 0), 500);
    }, 500);
  }, "atomics-wait-no-spurious-wakeup");
});

</script>
